Instruction message rework
Instead of sending a request for the party programme with every message the user sends. This is now requested during selection and is then saved/updated as a message when switching. This message is not rendered in the chat history due to its id = -99.

As soon as a PreStudy/PreTask page is created, this logic should be integrated into the next page button.