Created Zustaende
arguemnt:
A Zustnad that stores what the user has new learend about the manifesto;
Chebox:
A Zustand which stores if the user has accepted the information consent
Condition:
Stores if the user chats with an active (true) or an passive bot
Href:
Stores if the user has as task page the pdf or the chat system.
Informationgerthing:
Stores the prefered way of information gathering.
Likertscale:
Stores the selected values of the likertscales